requirements.txt
crewai>=0.28.0 
langchain>=0.1.0 
langchain-openai>=0.1.0 
openai>=1.0.0 
pandas>=2.0.0 
numpy>=1.24.0 
scikit-learn>=1.3.0 
netmiko>=4.2.0 
napalm>=4.1.0 
tensorflow>=2.13.0 
matplotlib>=3.7.0 
seaborn>=0.12.0 
python-dotenv>=1.0.0 
pydantic>=2.0.0 
requests>=2.31.0 
pytest>=7.4.0 
black>=23.0.0 
flake8>=6.0.0 
pyyaml>=6.0.0 
configparser>=6.0.0 
asyncio-mqtt>=0.13.0 
aiohttp>=3.8.0 
